requirements.txt
fastapi
uvicorn[standard]
numpy
joblib
tensorflow-cpu==2.15.0
pydantic
